+++ SWE1HttpServer/SWE1HttpServer/app/Models/Spell.cs
@@ -12,7 +12,7 @@ namespace SWE1HttpServer.app.Models
 {
     public class Spell : Card
     {
-        public Spell(ElementType element, int damage,string Id)
+        public Spell(ElementType element, float damage,string Id)
         {
             this.element = element;
             this.Damage = damage;
